Refactor DocumentServiceContext visible symbols method
Only the contexts which are subclasses of PositionedOperationContext
are caching the visible symbols for the cursor position.
