+++ language-server/modules/langserver-core/src/main/java/org/ballerinalang/langserver/contexts/PositionedOperationContextImpl.java
@@ -15,6 +15,7 @@
  */
 package org.ballerinalang.langserver.contexts;
 
+import io.ballerina.compiler.api.symbols.Symbol;
 import org.ballerinalang.langserver.LSContextOperation;
 import org.ballerinalang.langserver.commons.DocumentServiceContext;
 import org.ballerinalang.langserver.commons.LSOperation;
@@ -24,6 +25,8 @@ import org.ballerinalang.langserver.commons.workspace.WorkspaceManager;
 import org.eclipse.lsp4j.Position;
 import org.eclipse.lsp4j.jsonrpc.CancelChecker;
 
+import java.util.List;
+
 /**
  * Abstract implementation of the {@link PositionedOperationContext}.
  *
@@ -32,20 +35,59 @@ import org.eclipse.lsp4j.jsonrpc.CancelChecker;
 public abstract class PositionedOperationContextImpl extends AbstractDocumentServiceContext
         implements PositionedOperationContext {
 
-    @Deprecated(forRemoval = true)
-    PositionedOperationContextImpl(LSOperation operation,
-                                   String fileUri,
-                                   WorkspaceManager wsManager,
-                                   LanguageServerContext serverContext) {
-        super(operation, fileUri, wsManager, serverContext);
-    }
+    private final Position cursorPosition;
+
+    private int cursorPositionInTree = -1;
+    
+    private List<Symbol> visibleSymbols;
     
     PositionedOperationContextImpl(LSOperation operation,
                                    String fileUri,
+                                   Position cursorPosition,
                                    WorkspaceManager wsManager,
                                    LanguageServerContext serverContext,
                                    CancelChecker cancelChecker) {
         super(operation, fileUri, wsManager, serverContext, cancelChecker);
+        this.cursorPosition = cursorPosition;
+    }
+
+    /**
+     * {@inheritDoc}
+     * <p>
+     * Since we know the cursor position here, we can cache the visible symbols for the cursor position. This will be
+     * good for the performance since the chance of accessing the visible symbols of the cursor position repetitively is high.
+     *
+     * @param position Position at which visible symbols are needed
+     * @return Visible symbols
+     */
+    @Override
+    public List<Symbol> visibleSymbols(Position position) {
+        if (this.cursorPosition.equals(position)) {
+            if (this.visibleSymbols == null) {
+                this.visibleSymbols = super.visibleSymbols(position);
+            }
+            return this.visibleSymbols;
+        } else {
+            return super.visibleSymbols(position);
+        }
+    }
+
+    @Override
+    public void setCursorPositionInTree(int offset) {
+        if (this.cursorPositionInTree > -1) {
+            throw new RuntimeException("Setting the cursor offset more than once is not allowed");
+        }
+        this.cursorPositionInTree = offset;
+    }
+
+    @Override
+    public int getCursorPositionInTree() {
+        return this.cursorPositionInTree;
+    }
+
+    @Override
+    public Position getCursorPosition() {
+        return this.cursorPosition;
     }
 
     /**
